add cls to predict_system
