+++ tools/infer/text/README_CN.md
@@ -230,6 +230,50 @@ python deploy/eval_utils/eval_pipeline.py --gt_path path/to/gt.txt --pred_path p
 
 3、SVTR在混合精度模式下运行（amp_level=O2），因为它针对O2进行了优化。
 
+### 文本方向分类
+
+若图像中存在非正向的文字，可通过文本方向分类器对检测后的图像进行方向分类与矫正。若对输入图像运行文本方向分类与矫正，请执行
+```shell
+python tools/infer/text/predict_system.py --image_dir {path_to_img or dir_to_imgs} \
+                                          --det_algorithm DB++  \
+                                          --rec_algorithm CRNN  \
+                                          --cls_mode True
+```
+执行过程中，文本方向分类器将对文本检测所得图像列表进行方向分类，并对非正向的图像进行方向矫正。设置`--save_cls_result`为`True`可将文本方向分类结果保存至`{args.crop_res_save_dir}/cls_results.txt`中，其中`--crop_res_save_dir`是保存结果的目录，下面是一些结果的例子。
+
+- 文本方向分类
+
+<p align="center">
+  <img src="https://raw.githubusercontent.com/zhangjunlongtech/Material/refs/heads/main/CRNN_t1.png" width=150 />
+</p>
+<p align="center">
+  <em> word_01.png </em>
+</p>
+
+<p align="center">
+  <img src="https://raw.githubusercontent.com/zhangjunlongtech/Material/refs/heads/main/CRNN_t2.png" width=150 />
+</p>
+<p align="center">
+  <em> word_02.png </em>
+</p>
+
+分类结果：
+```text
+word_01.png   0     1.0
+word_02.png   180   1.0
+```
+可通过`--cls_algorithm`设置方向分类算法，并通过`--cls_amp_level`与`--cls_model_dir`来设置文本方向分类器的自动混合精度与权重文件。当前支持的文本方向分类网络为`mobilnet_v3`，已配置默认权重文件，该网络默认混合精度为`O0`，默认配置下方向分类支持0与180度，对于其他方向的分类我们将在未来予以支持。
+
+<center>
+
+  |**算法名称**|**网络名称**|**语言**|
+  | :------: | :------: | :------: |
+  | MV3 | mobilenet_v3 | 中/英|
+
+</center>
+
+有关更多参数说明和用法，请查看`tools/infer/text/config.py`
+
 ## 表格结构识别
 
 要对输入图像或包含多个图像的目录运行表格结构识别，请执行
@@ -310,56 +354,6 @@ HDL Cholesterol (mg/dL),42 ± 11.1,46 ± 11.4
 
 **注意事项：**
 1、如需更多参数说明和用法，请运行`python tools/infer/text/predict_table_recognition.py -h`或查看`tools/infer/text/config.py`
-
-## 文本方向分类器
-
-要对输入图像或包含多个图像的目录运行文本角度分类，请执行
-```shell
-python tools/infer/text/predict_cls.py  --image_dir {path_to_img or dir_to_imgs} --rec_algorithm MV3
-```
-运行后，推理结果保存在`{args.draw_img_save_dir}/cls_results.txt`中，其中`--draw_img_save_dir`是保存结果的目录，这是`./inference_results`的默认设置。下面是一些结果的例子。
-
-- 文本角度分类
-
-<p align="center">
-  <img src="https://raw.githubusercontent.com/zhangjunlongtech/Material/refs/heads/main/CRNN_t1.png" width=150 />
-</p>
-<p align="center">
-  <em> word_01.png </em>
-</p>
-
-<p align="center">
-  <img src="https://raw.githubusercontent.com/zhangjunlongtech/Material/refs/heads/main/CRNN_t2.png" width=150 />
-</p>
-<p align="center">
-  <em> word_02.png </em>
-</p>
-
-分类结果：
-```text
-word_01.png   0
-word_02.png   180
-```
-
-**注意事项：**
-- 有关更多参数说明和用法，请运行`python tools/infer/text/predict_cls.py -h`或查看`tools/infer/text/config.py`
-- 支持批量文本角度分类和单模文本角度分类。默认情况下启用批处理模式以提高速度。您可以通过`--cls_batch_num`设置批量大小。您还可以通过设置`--cls_batch_mode` False在单一图像模式下运行。
-- 可通过设置`--cls_model_dir` 来设置指定的权重文件，也可忽略此项设置，系统将加载默认权重文件。
-- 当前支持的角度分类器的网络为mobilenet_v3，可通过设置`--cls_algorithm` 来指定模型，其对应的`cls_amp_level`当前支持O0。
-
-### 支持的角度分类算法和网络
-
-<center>
-
-  |**算法名称**|**网络名称**|**语言**|
-  | :------: | :------: | :------: |
-  | MV3 | mobilenet_v3 | 中/英|
-
-</center>
-
-算法网络在`tools/infer/text/predict_cls.py`中定义
-
-目前，所列型号支持识别0与180度。对于其他角度的识别我们将很快予以支持。
 
 ## 参数列表
